+++ include/coldwarm.hpp
@@ -1,0 +1,6 @@
+#ifndef COLD_WARM_HPP
+#define COLD_WARM_HPP
+
+int getColdWarm(int argc, char* argv[]);
+
+#endif