+++ ChangeLog.md
@@ -11,3 +11,9 @@ All notable changes to this project will be documented in this file.
 ## 23/10/2024
 - Added the CSV parser header files and modified the makefile accordingly
 - Modified the cleaner/filter script to include all necessary dates
+
+## 24/10/2024
+- Created the code for extracting data for the coldest/warmest temperatures
+- Implemented the pre-processing bash script execution in the main executable
+- Implemented the choice of which data is desired in the main executable
+